backend/requirements.txt
fastapi~=0.119.0
uvicorn~=0.38.0
SQLAlchemy~=2.0.44
psycopg2-binary~=2.9.11
pydantic~=2.12.3
pydantic-settings~=2.11.0
nats-py~=2.11.0
